+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds `Placeholder Images` to the `Style Guide`.
+
 ## 4.1.6 (April 2, 2026)
 
 ### Updates